pyproject.toml
[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nethunt-mcp"
version = "0.1.0"
description = "Local MCP server for NetHunt CRM"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "WoodWay Expert (ТОВ \"Куросіо\")" }]
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.26.0,<4.0.0",
  "httpx>=0.28.1,<1.0.0",
  "mcp>=1.26.0,<2.0.0",
  "openai>=2.30.0,<3.0.0",
  "python-dotenv>=1.2.2,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2,<10.0.0",
  "pytest-asyncio>=1.3.0,<2.0.0",
  "respx>=0.22.0,<1.0.0",
]

[project.scripts]
nethunt-mcp = "nethunt_mcp.__main__:main"
nethunt-telegram-bot = "nethunt_mcp.telegram_main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
